requirements.txt
Django==5.2.4
django-ninja==1.3.0
django-ninja-jwt==5.3.5
uvicorn[standard]==0.30.6
gunicorn==23.0.0
psycopg2-binary
python-dotenv==1.0.1
stripe==10.11.0